Updated to current package structure.
