+++ .gitignore
@@ -175,3 +175,4 @@ buildNumber.properties
 .classpath
 
 # End of https://www.toptal.com/developers/gitignore/api/java,maven,intellij,git
+/.idea/sonarlint.xml
